% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_2y_.R
\name{grad_2y_}
\alias{grad_2y_}
\title{DREAM employment ratio months to DREAM benefit weeks}
\usage{
grad_2y_(grad_)
}
\arguments{
\item{grad_}{Character vector of DREAM employment ratio months (e.g. c("grad_2017_01","grad_2017_02",...))}
}
\value{
A character vector of DREAM benefit weeks (e.g. c("y_1701","y_1702",...)) that correspond to \code{yr}.
}
\description{
Generates a vector of DREAM benefit weeks that correspond to the input years.
}
