% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr2y_.R
\name{qtr2y_}
\alias{qtr2y_}
\title{Quaters to DREAM benefit weeks}
\usage{
qtr2y_(qtr)
}
\arguments{
\item{qtr}{Character vector of format yy'Q'q (e.g. c("17Q1","17Q2",...))}
}
\value{
A character vector of DREAM benefit weeks (e.g. c("y_1701","y_1702",...)) that correspond to \code{qtr}.
}
\description{
Generates a vector of DREAM benefit weeks that correspond to the input quarters.
}
