% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs}
\alias{runs}
\title{DREAM run length encoding}
\usage{
runs(data, row, period = NULL)
}
\arguments{
\item{data}{A \code{data.frame} (or \code{data.table}) containing a DREAM dataset.}

\item{row}{An integer valued row number (1 to \code{nrow}(data)).}

\item{period}{A character vector of DREAM weeks (e.g. c("y_1701","y_1702","y_1703",...)). If \code{NULL}, the default, all DREAM benefit weeks in \code{data} is used.}
}
\value{
A list with components:\cr
\code{lengths} An integer vector containing the length of each run.\cr
\code{values} A integer vector of same length as \code{lengths} with the corresponding values.\cr
\code{start} A character vector of same length as \code{lengths} with start weeks. \code{NA} imply that the run start prior to first week in \code{period}.\cr
\code{end} A character vector of same length as \code{lengths} with end weeks. \code{NA} imply that the run end after last week in \code{period}.
}
\description{
Computes the length, values, start, and end, of runs of equal values in a DREAM dataset.
}
