% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_2grad_.R
\name{y_2grad_}
\alias{y_2grad_}
\title{DREAM benefit weeks to employment rate months}
\usage{
y_2grad_(y_)
}
\arguments{
\item{y_}{Character vector of format 'y_'yyww (e.g. c("y_1701","y_1702",...))}
}
\value{
A character vector of employment rate months (e.g. c("grad_2017_01","grad_2017_02 ",...)) that correspond to \code{y_}.
}
\description{
Generates a vector of employment rate months that correspond to the input DREAM benefit weeks.
}
