% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2ak.R
\name{yr2ak}
\alias{yr2ak}
\title{Years to DREAM UI-fund membership}
\usage{
yr2ak(yr)
}
\arguments{
\item{yr}{Character vector of format yyyy (e.g. c("2015","2016",...))}
}
\value{
A character vector of DREAM UI-fund membership (e.g. c("ak15","ak16",...)) that correspond to \code{yr}.
}
\description{
Generates a vector of DREAM UI-fund membership that correspond to the input years.
}
