% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr2y_.R
\name{yr2y_}
\alias{yr2y_}
\title{Years to DREAM benefit weeks}
\usage{
yr2y_(yr)
}
\arguments{
\item{yr}{Character vector of format yyyy (e.g. c("2015","2016",...))}
}
\value{
A character vector of DREAM benefit weeks (e.g. c("y_1701","y_1702",...)) that correspond to \code{yr}.
}
\description{
Generates a vector of DREAM benefit weeks that correspond to the input years.
}
